+++ RELEASE_NOTES.md
@@ -55,6 +55,7 @@ The demo project contains a new `KeyboardKitSwiftUIPreviews` in which you can pr
 * `AutocompleteToolbar` has been renamed to `AutocompleteToolbarVew`.
 * `EmojiCategory.frequents` has been renamed to `frequent`.
 * `KeyboardActionHandler` now requires `canHandle(_:on:)` to be implemented.
+* `KeyboardAction` has new action types.
 * `KeyboardAction` has fewer `isXXX` properties.
 * `KeyboardAction` `.capsLock` and `shiftDown` are now part of `KeyboardAction.shift`.
 * `KeyboardActionRow.from` has been changed to an initializer.